+++ Models/SponsorEvent.cs
@@ -5,7 +5,7 @@ namespace GithubSponsorsWebhook.Models;
 public struct SponsorEvent
 {
     public GitHubAction Action { get; set; }
-    public long DatabaseId { get; set; }
+    public int DatabaseId { get; set; }
     public string LoginName { get; set; }
     public DateTime CreatedAt { get; set; }
     public GithubType GithubType { get; set; }
